Chapter 20 Challenges
Made table entries key on Values instead of ObjStrings, and moved the hash field to the Value struct.  Added separate helper functions to hash each of Bool, Nil, Number, and Obj.  The object hash is just a stub that returns zero, would have to figure out how many bytes were in the struct and loop over all the data with the FNV-1a hash function.